solvedbug with timeouts
git-svn-id: file:///semplestb/final_repo@1328 02200ff9-b5b2-46f0-9171-221b09c08c7b

+++ SemplestESB/src/semplest/server/ESB/AsyncServlet.java
@@ -78,7 +78,7 @@ public class AsyncServlet extends HttpServlet
 		{
 			try
 			{
-				defaultTimeoutMS =  Long.getLong(timeoutParam).longValue();
+				defaultTimeoutMS =  Long.getLong(timeoutParam);
 				timeout = true;
 			}
 			catch (NumberFormatException e)
